cleanup timeline related code
